// src/types/TaskType.ts
export enum TaskType {
  Housework = 'Hausarbeit z.B. putzen, waschen, einkaufen, Fenster putzen',
  Maintenance = 'Hausmeistertätigkeiten (z.B. Straße kehren, Schnee räumen)',
  Gardening = 'Gartenarbeit',
  AnimalCare = 'Versorgung von Tieren',
  EscortServices = 'Begleitdienste zum Arzt oder Spaziergänge',
  MealPreparation = 'Mahlzeiten zubereiten',
  Handicrafts = 'Handwerkliches (ohne spezielle Ausbildung)',
  ComputerSkills = 'Vermittlung von PC-Kenntnissen',
  Companionship = 'Gesellschaft leisten (Vorlesen, Gespräche, Spiele, Fernsehen usw.)'
}